requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
sqlparse>=0.4.4
httpx>=0.24.1
openenv-core>=0.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
